requirements.txt
beautifulsoup4==4.9.3
python-telegram-bot==13.3